refactor: 127-refactor default shipping address section
